feat/protectRoute

+++ emotors/src/modules/announcement/repositories/prisma/announcement-prisma.repository.ts
@@ -1,5 +1,5 @@
 /* eslint-disable prettier/prettier */
-import { Injectable } from '@nestjs/common/decorators';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { AnnouncementRepository } from '../announcement.repository';
 import { CreateAnnouncementDto } from '../../dto/create-announcement.dto';
 import { UpdateAnnouncementDto } from '../../dto/update-announcement.dto';
@@ -10,7 +10,10 @@ import { plainToInstance } from 'class-transformer';
 @Injectable()
 export class AnnouncementPrismaRepository implements AnnouncementRepository {
   constructor(private prisma: PrismaService) {}
-  async create(data: CreateAnnouncementDto): Promise<Announcement> {
+  async create(
+    data: CreateAnnouncementDto,
+    userId: string,
+  ): Promise<Announcement> {
     const announcement = new Announcement();
 
     Object.assign(announcement, { ...data });
@@ -29,31 +32,63 @@ export class AnnouncementPrismaRepository implements AnnouncementRepository {
         sellPrice: announcement.sellPrice,
         createdAt: announcement.createdAt || new Date(),
         updatedAt: announcement.updatedAt || undefined,
-        userId: announcement.userId,
+        userId,
       },
+      include: { user: true },
     });
 
     return plainToInstance(Announcement, newAnnouncement);
   }
   async findAll(): Promise<Announcement[]> {
-    const announcements = await this.prisma.announcement.findMany();
+    const announcements = await this.prisma.announcement.findMany({
+      include: { user: true },
+    });
     return plainToInstance(Announcement, announcements);
   }
   async findOne(id: string): Promise<Announcement> {
     const announcement = await this.prisma.announcement.findUnique({
       where: { id },
+      include: { user: true },
     });
     return plainToInstance(Announcement, announcement);
   }
-  async update(id: string, data: UpdateAnnouncementDto): Promise<Announcement> {
+  async update(
+    id: string,
+    data: UpdateAnnouncementDto,
+    userId: string,
+  ): Promise<Announcement> {
+    const user = await this.prisma.announcement.findFirst({
+      where: {
+        id,
+        userId,
+      },
+    });
+
+    if (!user) {
+      throw new UnauthorizedException('Unauthorized');
+    }
+
     const announcement = await this.prisma.announcement.update({
-      where: { id },
+      where: {
+        id: id,
+      },
+
       data: { ...data },
     });
 
     return plainToInstance(Announcement, announcement);
   }
-  async delete(id: string): Promise<void> {
+  async delete(id: string, userId: string): Promise<void> {
+    const user = await this.prisma.announcement.findFirst({
+      where: {
+        id,
+        userId,
+      },
+    });
+
+    if (!user) {
+      throw new UnauthorizedException('Unauthorized');
+    }
     await this.prisma.announcement.delete({
       where: { id },
     });
